+++ robocorp-code/src/robocorp_code/inspector/inspector_language_server.py
@@ -6,7 +6,6 @@ from typing import Literal, Optional
 from robocorp_ls_core.protocols import ActionResultDict  # type: ignore
 from robocorp_ls_core.robotframework_log import get_logger  # type: ignore
 
-from robocorp_code.inspector.common import LogLevel, log_call  # type: ignore
 
 log = get_logger(__name__)
 
@@ -22,14 +21,12 @@ class InspectorLanguageServer:
     def get_locators_json_path(self, directory: str) -> Path:
         return Path(directory) / "locators.json"
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_kill_inspectors(self, inspector: Optional[str]) -> None:
         inspector_api_client = self._inspector_server_manager.get_inspector_api_client()
         inspector_api_client.send_sync_message(
             "killInspectors", dict(inspector=inspector)
         )
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_manager_save_locator(
         self,
         locator: dict,
@@ -94,7 +91,6 @@ class InspectorLanguageServer:
         }
         return ret
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_manager_delete_locators(
         self,
         directory: str,
@@ -146,7 +142,6 @@ class InspectorLanguageServer:
         }
         return ret
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_manager_load_locators(self, directory: str) -> ActionResultDict:
         import json
 
@@ -199,7 +194,6 @@ class InspectorLanguageServer:
             }
             return ret
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_web_inspector_configure_browser(
         self, width: int = 1280, height: int = 720, url: str = ""
     ):
@@ -209,7 +203,6 @@ class InspectorLanguageServer:
             {"viewport_size": (width, height), "url": url},
         )
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_web_inspector_open_browser(self, url=None):
         inspector_api_client = self._inspector_server_manager.get_inspector_api_client()
         if url is None:
@@ -221,19 +214,16 @@ class InspectorLanguageServer:
 
         inspector_api_client.send_sync_message("openBrowser", dict(url=url, wait=True))
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_web_inspector_close_browser(self, **params) -> None:
         inspector_api_client = self._inspector_server_manager.get_inspector_api_client()
         inspector_api_client.send_sync_message("closeBrowser", {})
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_web_inspector_click(self, locator):
         inspector_api_client = self._inspector_server_manager.get_inspector_api_client()
         inspector_api_client.send_sync_message(
             "click", dict(locator=locator, wait=True)
         )
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_web_inspector_start_pick(self, **params):
         from robocorp_ls_core import uris
 
@@ -249,12 +239,10 @@ class InspectorLanguageServer:
             "startPick", {"url_if_new": url, "wait": True}
         )
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_web_inspector_stop_pick(self, **params):
         inspector_api_client = self._inspector_server_manager.get_inspector_api_client()
         inspector_api_client.send_sync_message("stopPick", {"wait": True})
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_web_inspector_validate_locator(self, **params):
         inspector_api_client = self._inspector_server_manager.get_inspector_api_client()
         return partial(
@@ -263,7 +251,6 @@ class InspectorLanguageServer:
             {"locator": params["locator"], "url": params["url"], "wait": True},
         )
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_windows_inspector_parse_locator(self, locator: str):
         inspector_api_client = self._inspector_server_manager.get_inspector_api_client()
         # Not blocking (return callback to run in thread).
@@ -273,7 +260,6 @@ class InspectorLanguageServer:
             {"locator": locator},
         )
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_windows_inspector_set_window_locator(self, locator: str):
         inspector_api_client = self._inspector_server_manager.get_inspector_api_client()
         # Not blocking (return callback to run in thread).
@@ -283,7 +269,6 @@ class InspectorLanguageServer:
             {"locator": locator},
         )
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_windows_inspector_list_windows(self):
         inspector_api_client = self._inspector_server_manager.get_inspector_api_client()
         # Not blocking (return callback to run in thread).
@@ -293,7 +278,6 @@ class InspectorLanguageServer:
             {},
         )
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_windows_inspector_start_pick(self):
         inspector_api_client = self._inspector_server_manager.get_inspector_api_client()
         # Not blocking (return callback to run in thread).
@@ -303,7 +287,6 @@ class InspectorLanguageServer:
             {},
         )
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_windows_inspector_stop_pick(self):
         inspector_api_client = self._inspector_server_manager.get_inspector_api_client()
         # Not blocking (return callback to run in thread).
@@ -313,7 +296,6 @@ class InspectorLanguageServer:
             {},
         )
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_windows_inspector_start_highlight(
         self,
         locator: str,
@@ -332,7 +314,6 @@ class InspectorLanguageServer:
             ),
         )
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_windows_inspector_collect_tree(
         self,
         locator: str,
@@ -351,7 +332,6 @@ class InspectorLanguageServer:
             ),
         )
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_windows_inspector_stop_highlight(self):
         inspector_api_client = self._inspector_server_manager.get_inspector_api_client()
         # Not blocking (return callback to run in thread).
@@ -361,7 +341,6 @@ class InspectorLanguageServer:
             {},
         )
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_image_inspector_start_pick(
         self, minimize: Optional[bool], confidence_level: Optional[int]
     ):
@@ -372,12 +351,10 @@ class InspectorLanguageServer:
             {"minimize": minimize, "confidence_level": confidence_level},
         )
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_image_inspector_stop_pick(self):
         inspector_api_client = self._inspector_server_manager.get_inspector_api_client()
         return partial(inspector_api_client.send_sync_message, "imageStopPick", {})
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_image_inspector_validate_locator(
         self, locator: dict, confidence_level: Optional[bool]
     ):
@@ -388,7 +365,6 @@ class InspectorLanguageServer:
             {"locator": locator, "confidence_level": confidence_level},
         )
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_image_inspector_save_image(self, root_directory: str, image_base64: str):
         inspector_api_client = self._inspector_server_manager.get_inspector_api_client()
         return partial(
@@ -397,7 +373,6 @@ class InspectorLanguageServer:
             {"root_directory": root_directory, "image_base64": image_base64},
         )
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_java_inspector_parse_locator(self, locator: str):
         inspector_api_client = self._inspector_server_manager.get_inspector_api_client()
         # Not blocking (return callback to run in thread).
@@ -407,7 +382,6 @@ class InspectorLanguageServer:
             {"locator": locator},
         )
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_java_inspector_set_window_locator(self, locator: str):
         inspector_api_client = self._inspector_server_manager.get_inspector_api_client()
         # Not blocking (return callback to run in thread).
@@ -417,7 +391,6 @@ class InspectorLanguageServer:
             {"locator": locator},
         )
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_java_inspector_list_windows(self):
         inspector_api_client = self._inspector_server_manager.get_inspector_api_client()
         # Not blocking (return callback to run in thread).
@@ -427,7 +400,6 @@ class InspectorLanguageServer:
             {},
         )
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_java_inspector_start_pick(self):
         inspector_api_client = self._inspector_server_manager.get_inspector_api_client()
         # Not blocking (return callback to run in thread).
@@ -437,7 +409,6 @@ class InspectorLanguageServer:
             {},
         )
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_java_inspector_stop_pick(self):
         inspector_api_client = self._inspector_server_manager.get_inspector_api_client()
         # Not blocking (return callback to run in thread).
@@ -447,7 +418,6 @@ class InspectorLanguageServer:
             {},
         )
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_java_inspector_start_highlight(
         self,
         locator: str,
@@ -464,7 +434,6 @@ class InspectorLanguageServer:
             ),
         )
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_java_inspector_collect_tree(
         self,
         locator: str,
@@ -481,7 +450,6 @@ class InspectorLanguageServer:
             ),
         )
 
-    @log_call(log_level=LogLevel.DEBUG)
     def m_java_inspector_stop_highlight(self):
         inspector_api_client = self._inspector_server_manager.get_inspector_api_client()
         # Not blocking (return callback to run in thread).